api/requirements/dev.txt
Django==2.2.9

psycopg2==2.7.6.1

psycopg2-binary==2.7.6.1

django-dotenv==1.4.2

ipython

djangorestframework==3.9.0

djangorestframework-jwt==1.11.0

simplejson==3.16.0

django-cors-headers==2.4.0

django-redis==4.9.0
redis==2.10.6

celery==4.2.1

jsonschema==2.6.0

python-alipay-sdk
Crypto
pycryptodomex

requests

lxml

django-filter==2.2.0

bs4

django-safedelete==0.5.4

